Redstone armor should drop redstone essence which should last for a couple of seconds activating all redstone devices.
