+++ src/main/java/de/melanx/MoreVanillaArmor/MoreVanillaArmor.java
@@ -1,13 +1,16 @@
 package de.melanx.MoreVanillaArmor;
 
 //import de.melanx.MoreVanillaArmor.init.ModEffects;
+import de.melanx.MoreVanillaArmor.blocks.ModBlocks;
 import de.melanx.MoreVanillaArmor.util.CreativeTab;
 import de.melanx.MoreVanillaArmor.util.Registry;
 import net.minecraft.item.ItemGroup;
 import net.minecraftforge.common.MinecraftForge;
 //import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 //import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -25,9 +28,12 @@ public class MoreVanillaArmor {
         instance = this;
 
         MinecraftForge.EVENT_BUS.register(this);
+
+        final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+        ModBlocks.BLOCKS.register(modEventBus);
+
         Registry.registerArmor();
         Registry.registerArmorEffects();
-        Registry.registerBlocks();
     }
 
 }
